# anchore_engine/services/catalog/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Catalog service for anchore-engine"
  version: "0.9.2"
  title: "Catalog"
  contact:
    email: "dev@anchore.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
#host: "localhost:83"
basePath: "/"
...
schemes:
- "https"
- "http"
paths:
...
  /imports/images:
    post:
      x-swagger-router-controller: anchore_engine.services.catalog.api.controllers.imports
      operationId: create_operation
      x-anchore-authz-action: importImage
      summary: "Begin the import of an image analyzed by Syft into the system"
      produces:
        - "application/json"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportOperation"
        500:
          description: "Internal Error"
    get:
      x-swagger-router-controller: anchore_engine.services.catalog.api.controllers.imports
      operationId: list_operations
      x-anchore-authz-action: importImage
      summary: "Lists in-progress imports"
      produces:
        - "application/json"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImports"
        500:
          description: "Internal Error"
  /imports/images/{operation_id}:
    get:
      x-swagger-router-controller: anchore_engine.services.catalog.api.controllers.imports
      operationId: get_operation
      x-anchore-authz-action: importImage
      summary: "Get detail on a single import"
      produces:
        - "application/json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportOperation"
        500:
          description: "Internal Error"
    delete:
      x-swagger-router-controller: anchore_engine.services.catalog.api.controllers.imports
      operationId: invalidate_operation
      x-anchore-authz-action: importImage
      summary: "Invalidate operation ID so it can be garbage collected"
      produces:
        - "application/json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportOperation"
        500:
          description: "Internal Error"
    put:
      x-swagger-router-controller: anchore_engine.services.catalog.api.controllers.imports
      operationId: update_operation
      x-anchore-authz-action: null
      summary: "Internal operation to update import state"
      produces:
        - "application/json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
        - name: operation
          in: body
          required: true
          schema:
            $ref: "#/definitions/ImageImportOperation"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportOperation"
        500:
          description: "Internal Error"
  /imports/images/{operation_id}/packages:
    get:
      x-swagger-router-controller: anchore_engine.services.catalog.api.controllers.imports
      operationId: list_import_packages
      x-anchore-authz-action: importImage
      summary: "Begin the import of an image analyzed by Syft into the system"
      produces:
        - "application/json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/PackageDigestList"
        500:
          description: "Internal Error"
    post:
      x-swagger-router-controller: anchore_engine.services.catalog.api.controllers.imports
      operationId: import_image_packages
      x-anchore-authz-action: importImage
      summary: "Begin the import of an image analyzed by Syft into the system"
      produces:
        - "application/json"
      consumes:
        - "application/json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
        - name: sbom
          in: body
          required: true
          schema:
            $ref: "#/definitions/ImagePackageManifest"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportContentResponse"
        500:
          description: "Internal Error"
  /imports/images/{operation_id}/dockerfile:
    post:
      x-swagger-router-controller: anchore_engine.services.catalog.api.controllers.imports
      operationId: import_image_dockerfile
      x-anchore-authz-action: importImage
      summary: "Begin the import of an image analyzed by Syft into the system"
      produces:
        - "application/json"
      consumes:
        - "text/plain; utf-8"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportContentResponse"
        500:
          description: "Internal Error"
  /imports/images/{operation_id}/manifest:
    post:
      x-swagger-router-controller: anchore_engine.services.catalog.api.controllers.imports
      operationId: import_image_manifest
      x-anchore-authz-action: importImage
      summary: "Import a docker or OCI distribution manifest to associate with the image"
      produces:
        - "application/json"
      consumes:
        - "application/vnd.oci.image.manifest.v1+json"
        - "application/vnd.docker.distribution.manifest.v2+json"
        - "application/vnd.docker.distribution.manifest.v1+json" # Older Docker spec
        - "application/vnd.docker.distribution.manifest.v1+prettyjws" # Signed version
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportContentResponse"
        500:
          description: "Internal Error"
  /imports/images/{operation_id}/parent_manifest:
    post:
      x-swagger-router-controller: anchore_engine.services.catalog.api.controllers.imports
      operationId: import_image_parent_manifest
      x-anchore-authz-action: importImage
      summary: "Import a docker or OCI distribution manifest list to associate with the image"
      produces:
        - "application/json"
      consumes:
        - "application/vnd.docker.distribution.manifest.list.v2+json"
        - "application/vnd.oci.image.index.v1+json"
      parameters:
        - name: operation_id
          in: path
          type: string
          required: true
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ImageImportContentResponse"
        500:
          description: "Internal Error"
definitions:
...
  ImageImportOperation:
    type: object
    description: An import record, creating a unique identifier for referencing the operation as well as its state
    properties:
      uuid:
        type: string
      status:
        type: string
        enum:
          - pending
          - queued
          - processing
          - complete
          - failed
          - expired
      expires_at:
        type: string
        format: date-time
      created_at:
        type: string
        format: date-time
  ImageImportContentResponse:
    type: object
    properties:
      digest:
        type: string
      created_at:
        type: string
        format: date-time
...
  ImageImports:
    type: array
    items:
      $ref: "#/definitions/ImageImportOperation"
  ImagePackageManifest:
    type: object
    # TODO: Insert Syft schema v1 here in PackageManifest
  PackageDigestList:
    type: array
    items:
      type: string
      description: String digest of an uploaded content